rename hook to useResourceOwner
